rel/gae_standard_fileupload/requirements.txt
Flask==2.0.1
google-api-python-client==2.12.0
google-auth==1.31.0
google-cloud-storage==1.38.0
pyjwt[crypto]>=1.7.1
